requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-import-export==2.0.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
